Version 0.1
Character creation and introduction are finished. Also the basics of the game loop are down as well.

I will work on the combat section and the conditions for beating the game next.

I plan to add a start screen with an optional tutorial and a way to skip the intro.